+++ lib/components/search_bar.dart
