+++ lib/components/time_picker.dart
@@ -1,19 +1,14 @@
 import 'package:easy_localization/easy_localization.dart';
 import 'package:flutter/material.dart';
+import '../ui/themes.dart';
 
-import '../themes.dart';
+///Személyre szabott időpontválasztó widget.
+class TimePicker extends StatefulWidget {
+  ///A megváltozott időpont jelzésére szolgáló callback
+  final ValueChanged<TimeOfDay> onChanged;
 
-///A [DatePicker] widgetben kiválasztható legkorábbi és legkésőbbi dátum
-///különbsége az aktuális időponttól.
-const kDatePickerDifference = Duration(days: 730);
-
-///Személyre szabott dátumválasztó widget.
-class DatePicker extends StatefulWidget {
-  ///A megváltozott dátum jelzésére szolgáló callback
-  final ValueChanged<DateTime> onChanged;
-
-  ///Eredeti dátum
-  final DateTime date;
+  ///Eredeti időpont
+  final TimeOfDay time;
 
   ///Szöveg színe
   final Color color;
@@ -21,31 +16,30 @@ class DatePicker extends StatefulWidget {
   ///Szöveg mérete
   final double fontSize;
 
-  const DatePicker(
+  const TimePicker(
       {Key? key,
-      required this.date,
+      required this.time,
       required this.onChanged,
       this.color = szikTarawera,
       this.fontSize = 14})
       : super(key: key);
 
   @override
-  _DatePickerState createState() => _DatePickerState();
+  _TimePickerState createState() => _TimePickerState();
 }
 
-class _DatePickerState extends State<DatePicker> {
-  void _selectDate() async {
-    final newDate = await showDatePicker(
+class _TimePickerState extends State<TimePicker> {
+  void _selectTime() async {
+    final newTime = await showTimePicker(
       context: context,
-      initialDate: widget.date,
-      firstDate: DateTime.now().subtract(kDatePickerDifference),
-      lastDate: DateTime.now().add(kDatePickerDifference),
+      initialTime: widget.time,
+      initialEntryMode: TimePickerEntryMode.dial,
       confirmText: 'BUTTON_OK'.tr().toUpperCase(),
       cancelText: 'BUTTON_CANCEL'.tr().toUpperCase(),
-      helpText: 'DATE_PICKER_HELP'.tr().toUpperCase(),
+      helpText: 'TIME_PICKER_HELP'.tr().toUpperCase(),
     );
-    if (newDate != null) {
-      widget.onChanged(newDate);
+    if (newTime != null) {
+      widget.onChanged(newTime);
     }
   }
 
@@ -61,9 +55,9 @@ class _DatePickerState extends State<DatePicker> {
       ),
       padding: const EdgeInsets.fromLTRB(5, 10, 5, 10),
       child: GestureDetector(
-        onTap: _selectDate,
+        onTap: _selectTime,
         child: Text(
-          DateFormat('yyyy. MM. dd.').format(widget.date),
+          widget.time.format(context),
           style: Theme.of(context).textTheme.button!.copyWith(
               color: widget.color,
               fontSize: widget.fontSize,
