+++ lib/screens/error_screen.dart
@@ -1,5 +1,5 @@
 import 'package:flutter/material.dart';
-import '../../main.dart';
+import '../main.dart';
 
 class ErrorScreenArguments {
   Object error;
