+++ lib/screens/janitor_edit_admin.dart
@@ -1,11 +1,11 @@
 import 'package:easy_localization/easy_localization.dart';
 import 'package:flutter/material.dart';
 
-import '../../business/janitor_manager.dart';
-import '../../main.dart';
-import '../../models/tasks.dart';
-import '../widgets/alert_dialog.dart';
-import '../widgets/searchable_options.dart';
+import '../business/janitor_manager.dart';
+import '../components/alert_dialog.dart';
+import '../components/searchable_options.dart';
+import '../main.dart';
+import '../models/tasks.dart';
 
 class JanitorEditAdminArguments {
   JanitorTask task;
