+++ lib/screens/profile_screen.dart
@@ -1,10 +1,10 @@
 import 'package:flutter/material.dart';
 
-class ProfilePage extends StatelessWidget {
+class ProfileScreen extends StatelessWidget {
   static const String route = '/profile';
   static const String shortRoute = '/me';
 
-  const ProfilePage({Key key = const Key('ProfilePage')}) : super(key: key);
+  const ProfileScreen({Key key = const Key('ProfileScreen')}) : super(key: key);
 
   @override
   Widget build(BuildContext context) {
