+++ lib/screens/reservation_details.dart
@@ -1,17 +1,12 @@
 import 'package:easy_localization/easy_localization.dart';
 import 'package:flutter/material.dart';
 
-import '../../business/reservation_manager.dart';
+import '../business/reservation_manager.dart';
 import 'error_screen.dart';
-
-class ReservationDetailsArguments {
-  String title;
-
-  ReservationDetailsArguments({required this.title});
-}
 
 class ReservationDetailsScreen extends StatefulWidget {
   static const String route = '/reservation/details';
+
   final String title;
   const ReservationDetailsScreen({Key? key, required this.title})
       : super(key: key);
@@ -33,18 +28,19 @@ class _ReservationDetailsScreenState extends State<ReservationDetailsScreen> {
   @override
   Widget build(BuildContext context) {
     return FutureBuilder<void>(
-        future: reservation.refresh(),
-        builder: (context, snapshot) {
-          if (snapshot.connectionState == ConnectionState.waiting) {
-            //Shrimmer
-            return const Scaffold();
-          } else if (snapshot.hasError) {
-            return ErrorScreen(error: snapshot.error ?? 'ERROR_UNKNOWN'.tr());
-          } else {
-            return const Scaffold(
-                //Ide
-                );
-          }
-        });
+      future: reservation.refresh(),
+      builder: (context, snapshot) {
+        if (snapshot.connectionState == ConnectionState.waiting) {
+          //Shrimmer
+          return const Scaffold();
+        } else if (snapshot.hasError) {
+          return ErrorScreen(error: snapshot.error ?? 'ERROR_UNKNOWN'.tr());
+        } else {
+          return const Scaffold(
+              //Ide
+              );
+        }
+      },
+    );
   }
 }
